Extract CardFields into card_fields.dart
