+++ lib/src/sdk/cards/card_fields.dart
@@ -1,0 +1,31 @@
+enum CardFields {
+  all,
+  id,
+  badges,
+  checkItemStates,
+  closed,
+  dateLastActivity,
+  desc,
+  descData,
+  due,
+  dueComplete,
+  idAttachmentCover,
+  idBoard,
+  idChecklists,
+  idLabels,
+  idList,
+  idMembers,
+  idMembersVoted,
+  idShort,
+  labels,
+  manualCoverAttachment,
+  name,
+  pos,
+  shortLink,
+  shortUrl,
+  subscribed,
+  url,
+  address,
+  locationName,
+  coordinates,
+}
