+++ lib/src/sdk/cards/cards.dart
@@ -3,6 +3,7 @@ export 'attachment_id.dart';
 export 'card.dart';
 export 'card_badges.dart';
 export 'card_client.dart';
+export 'card_fields.dart';
 export 'card_id.dart';
 export 'card_label.dart';
 export 'card_models.dart';
