+++ lib/src/sdk/cards/trello_card.dart
@@ -1,9 +1,9 @@
 import '../coordinates.dart' show Coordinates;
 import '../trello_object.dart' show TrelloObject;
 import 'card_badges.dart' show CardBadges;
+import 'card_fields.dart' show CardFields;
 import 'card_id.dart' show CardId;
 import 'card_label.dart' show CardLabel;
-import 'card_models.dart' show CardBadges, CardFields;
 
 class TrelloCard extends TrelloObject<CardFields> {
   TrelloCard(source, List<CardFields> fields)
